requirements.txt
sanic
python-dotenv
pandas
sklearn
xgboost
